+++ packages/core/src/storage/options.js
@@ -45,6 +45,7 @@ export function createOptions(plugins) {
             hour: 'numeric',
             minute: '2-digit'
         },
+        filterEventsWithResources: false,
         filterResourcesWithEvents: false,
         firstDay: 0,
         flexibleSlotTimeLimits: false,  // ec option